Fix quickaddnew sometimes getting entwined twice
Issue stems from jQuery plugins like Select2 re-writing the dom and
causing entwine to match quickaddnew to the field a second time.
This then causes issues with doubled up buttons and dialogs.
The other potential fix is making sure quickaddnew fires last in
the page build but this can't be guaranteed.